feat(cinema-frontend) : profile side done -if user not logged in, program asks user do you want to login. if user says no, returns back if user says yes, moved login screen. If login is successful, return back to main window -all files moved to initial position of cinema-frontend file
