+++ .gitignore
@@ -1,1 +1,1 @@
-cinema-frontend-init/node_modules
+node_modules
