--- cinema-frontend-init/pages/islogin.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.css">
-<body>
-    <p>Do you want to log in?</p>
-    <button type="button" id="yes-button" class="btn btn-primary btn-lg btn-block" onclick="openLogin()">Yes</button>
-    <button type="button" id="no-button" class="btn btn-secondary btn-lg btn-block" onclick="closeScreen()">No</button>
-</body>
-<script>
-    function openLogin() {
-       closeScreen();
-       function openLoginScreen(){
-            let window = require('electron').BrowserWindow;
-            window.open('https://www.google.com');
-        }
-    }
-    function closeScreen(){
-        window.open('', '_self', '');
-        window.close();
-    }
-</script>
-</html>
