+++ js/alert.js
@@ -6,7 +6,15 @@ function moveToChooseCinema() {
     localStorage.setItem("password", password);
     
     console.log(username + " " + password);
-    return username === "melihcanclk" && password === "123";
+    if(username === "melihcanclk" && password === "123"){
+        localStorage.setItem('isLogin', 'true')
+        window.open('', '_self', '');
+        window.close();
+        console.log(localStorage.getItem('islogin'))
+        return true;
+    }else{
+        return false;
+    }
     //if returns true, moves second page which make us choose cinema saloons but if
     // returns false, make us perform inside this function
 }
