+++ pages/islogin.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.css">
+<body>
+    <p>Do you want to log in?</p>
+    <button type="button" id="yes-button" class="btn btn-primary btn-lg btn-block" onclick="javascript:openLoginPage();">Yes</button>
+    <button type="button" id="no-button" class="btn btn-secondary btn-lg btn-block" onclick="javascript:closeScreen();">No</button>
+    <script src="../js/openLogin.js"></script>
+</body>
+</html>
