+++ pages/loginScreen.html
@@ -12,7 +12,7 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
     <img class="image" src="../assets/burak.png" alt="∞">
     <div id="input-div">
-        <form action="choosingCinema.html" onsubmit=" return moveToChooseCinema() ;">
+        <form action="#" onsubmit=" return moveToChooseCinema() ;">
             <div>
                 <label for="username-input">Username</label>
                 <input type="text" class="form-control" name="username" id="username-input" placeholder="Enter username">
